backend/pyproject.toml
[tool.poetry]
name="InsightGrid"
version="0.1.0"
description="InsightGrid API"
authors = [
    "Kavin <kavin6462@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.103.0"
uvicorn = "^0.20.0"

pydantic = "^2.5.0"
pydantic_settings = "^2.2.0"

SQLAlchemy = "^2.0.0"
redis = "^5.0.0"

[tool.poetry.dev-dependencies]
isort="^5.13.0"
black=  "^24.4.0"
flake8= "^7.0.0"

[tool.isort]
profile = "black"
src_paths = ["app"]

[tool.black]
line-length = 88
target-version = ['py310','py311']


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
